# api/OpenAPI.yaml
openapi: 3.0.3
info:
  title: enman API
  version: 0.1.0
servers:
  - description: Development server
    url: http://localhost:1099
  - description: Production server
    url: https://your-domain.example.com
...
paths:
...
  /auth/register:
    post:
      summary: ユーザー登録
      operationId: postAuthRegister
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /auth/login:
    post:
      summary: ログイン
      operationId: postAuthLogin
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /auth/refresh:
    post:
      summary: アクセストークン更新
      operationId: postAuthRefresh
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: トークン更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /auth/logout:
    delete:
      summary: ログアウト
      operationId: deleteAuthLogout
      tags:
        - auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: ログアウト成功
        default:
          $ref: "./components/schemas/response/default.yaml"
  /families:
    post:
      summary: 家族グループ作成
      operationId: postFamilies
      tags:
        - families
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyRequest"
      responses:
        "201":
          description: 家族グループ作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /families/join:
    post:
      summary: 招待コードで家族グループに参加
      operationId: postFamiliesJoin
      tags:
        - families
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinFamilyRequest"
      responses:
        "200":
          description: 参加成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /tasks:
    get:
      summary: タスク一覧取得
      operationId: getTasks
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: "pending: これからやること / today_done: 今日終わったこと"
          schema:
            type: string
            enum: [pending, today_done]
      responses:
        "200":
          description: タスク一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
    post:
      summary: タスク作成
      operationId: postTasks
      tags:
        - tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: タスク作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /tasks/{taskId}:
    patch:
      summary: タスク編集
      operationId: patchTask
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: タスク編集成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
    delete:
      summary: タスク削除
      operationId: deleteTask
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: タスク削除成功
        default:
          $ref: "./components/schemas/response/default.yaml"
  /tasks/{taskId}/done:
    patch:
      summary: タスク完了報告
      operationId: patchTasksDone
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 完了報告成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /task-templates:
    get:
      summary: 家事テンプレート一覧取得
      operationId: getTaskTemplates
      tags:
        - task-templates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: テンプレート一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskTemplateResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /tasks/{taskId}/appreciations:
    post:
      summary: ありがとうスタンプ送信
      operationId: postTasksAppreciation
      tags:
        - appreciations
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAppreciationRequest"
      responses:
        "201":
          description: スタンプ送信成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppreciationResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /notifications:
    get:
      summary: 通知一覧取得（自分宛の感謝）
      operationId: getNotifications
      tags:
        - notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 通知一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppreciationResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /stats/contributions:
    get:
      summary: 家族の貢献度統計取得
      operationId: getStatsContributions
      tags:
        - stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 貢献度統計取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContributionItem"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /family/timeline:
    get:
      summary: 感謝タイムライン取得（家族全体）
      operationId: getFamilyTimeline
      tags:
        - appreciations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: タイムライン取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppreciationResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /families/goals:
    get:
      summary: ご褒美目標一覧取得
      operationId: getFamiliesGoals
      tags:
        - goals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 目標一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyGoalResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
    post:
      summary: ご褒美目標作成
      operationId: postFamiliesGoals
      tags:
        - goals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyGoalRequest"
      responses:
        "201":
          description: 目標作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyGoalResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /recurrence-rules:
    get:
      summary: 繰り返しルール一覧取得
      operationId: getRecurrenceRules
      tags:
        - recurrence-rules
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurrenceRuleResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
    post:
      summary: 繰り返しルール作成
      operationId: postRecurrenceRules
      tags:
        - recurrence-rules
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceRuleRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurrenceRuleResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
  /recurrence-rules/{ruleId}:
    patch:
      summary: 繰り返しルール更新
      operationId: patchRecurrenceRules
      tags:
        - recurrence-rules
      security:
        - bearerAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceRuleRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurrenceRuleResponse"
        default:
          $ref: "./components/schemas/response/default.yaml"
    delete:
      summary: 繰り返しルール削除（以降のインスタンスをすべて終了）
      operationId: deleteRecurrenceRule
      tags:
        - recurrence-rules
      security:
        - bearerAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 削除成功
        default:
          $ref: "./components/schemas/response/default.yaml"
  /calendar:
    get:
      summary: 月別カレンダー取得（繰り返しタスク展開済み）
      operationId: getCalendar
      tags:
        - calendar
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 5
      responses:
        "200":
          description: カレンダー取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDayItem"
        default:
          $ref: "./components/schemas/response/default.yaml"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          example: "田中太郎"
        email:
          type: string
          format: email
          example: "taro@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
        role:
          type: string
          example: "パパ"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    CreateFamilyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "田中家"
    JoinFamilyRequest:
      type: object
      required:
        - invite_code
      properties:
        invite_code:
          type: string
          example: "ABC123DEF456"
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        access_token:
          type: string
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...
    UpdateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "皿洗い"
        category:
          type: string
          nullable: true
          example: "キッチン"
...
    FamilyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        invite_code:
          type: string
        created_at:
          type: string
          format: date-time
...