# backend/internal/interface/openapi/api_server.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package openapi

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ログイン
	// (POST /auth/login)
	PostAuthLogin(ctx echo.Context) error
	// ログアウト
	// (DELETE /auth/logout)
	DeleteAuthLogout(ctx echo.Context) error
	// アクセストークン更新
	// (POST /auth/refresh)
	PostAuthRefresh(ctx echo.Context) error
	// ユーザー登録
	// (POST /auth/register)
	PostAuthRegister(ctx echo.Context) error
	// 月別カレンダー取得（繰り返しタスク展開済み）
	// (GET /calendar)
	GetCalendar(ctx echo.Context, params GetCalendarParams) error
	// 家族グループ作成
	// (POST /families)
	PostFamilies(ctx echo.Context) error
	// ご褒美目標一覧取得
	// (GET /families/goals)
	GetFamiliesGoals(ctx echo.Context) error
	// ご褒美目標作成
	// (POST /families/goals)
	PostFamiliesGoals(ctx echo.Context) error
	// 招待コードで家族グループに参加
	// (POST /families/join)
	PostFamiliesJoin(ctx echo.Context) error
	// 感謝タイムライン取得（家族全体）
	// (GET /family/timeline)
	GetFamilyTimeline(ctx echo.Context) error
	// Health check
	// (GET /health)
	GetHealth(ctx echo.Context) error
	// 通知一覧取得（自分宛の感謝）
	// (GET /notifications)
	GetNotifications(ctx echo.Context) error
	// 繰り返しルール一覧取得
	// (GET /recurrence-rules)
	GetRecurrenceRules(ctx echo.Context) error
	// 繰り返しルール作成
	// (POST /recurrence-rules)
	PostRecurrenceRules(ctx echo.Context) error
	// 繰り返しルール削除（以降のインスタンスをすべて終了）
	// (DELETE /recurrence-rules/{ruleId})
	DeleteRecurrenceRule(ctx echo.Context, ruleId int64) error
	// 繰り返しルール更新
	// (PATCH /recurrence-rules/{ruleId})
	PatchRecurrenceRules(ctx echo.Context, ruleId int64) error
	// 家族の貢献度統計取得
	// (GET /stats/contributions)
	GetStatsContributions(ctx echo.Context) error
	// 家事テンプレート一覧取得
	// (GET /task-templates)
	GetTaskTemplates(ctx echo.Context) error
	// タスク一覧取得
	// (GET /tasks)
	GetTasks(ctx echo.Context, params GetTasksParams) error
	// タスク作成
	// (POST /tasks)
	PostTasks(ctx echo.Context) error
	// タスク削除
	// (DELETE /tasks/{taskId})
	DeleteTask(ctx echo.Context, taskId int64) error
	// タスク編集
	// (PATCH /tasks/{taskId})
	PatchTask(ctx echo.Context, taskId int64) error
	// ありがとうスタンプ送信
	// (POST /tasks/{taskId}/appreciations)
	PostTasksAppreciation(ctx echo.Context, taskId int64) error
	// タスク完了報告
	// (PATCH /tasks/{taskId}/done)
	PatchTasksDone(ctx echo.Context, taskId int64) error
	// 自分のプロフィール取得
	// (GET /users/me)
	GetUsersMe(ctx echo.Context) error
	// プロフィール更新
	// (PATCH /users/me)
	PatchUsersMe(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostAuthLogin converts echo context to params.
func (w *ServerInterfaceWrapper) PostAuthLogin(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAuthLogin(ctx)
	return err
}

// DeleteAuthLogout converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteAuthLogout(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteAuthLogout(ctx)
	return err
}

// PostAuthRefresh converts echo context to params.
func (w *ServerInterfaceWrapper) PostAuthRefresh(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAuthRefresh(ctx)
	return err
}

// PostAuthRegister converts echo context to params.
func (w *ServerInterfaceWrapper) PostAuthRegister(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAuthRegister(ctx)
	return err
}

// GetCalendar converts echo context to params.
func (w *ServerInterfaceWrapper) GetCalendar(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetCalendarParams
	// ------------- Required query parameter "year" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "year", ctx.QueryParams(), &params.Year, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter year: %s", err))
	}

	// ------------- Required query parameter "month" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "month", ctx.QueryParams(), &params.Month, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter month: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetCalendar(ctx, params)
	return err
}

// PostFamilies converts echo context to params.
func (w *ServerInterfaceWrapper) PostFamilies(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostFamilies(ctx)
	return err
}

// GetFamiliesGoals converts echo context to params.
func (w *ServerInterfaceWrapper) GetFamiliesGoals(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetFamiliesGoals(ctx)
	return err
}

// PostFamiliesGoals converts echo context to params.
func (w *ServerInterfaceWrapper) PostFamiliesGoals(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostFamiliesGoals(ctx)
	return err
}

// PostFamiliesJoin converts echo context to params.
func (w *ServerInterfaceWrapper) PostFamiliesJoin(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostFamiliesJoin(ctx)
	return err
}

// GetFamilyTimeline converts echo context to params.
func (w *ServerInterfaceWrapper) GetFamilyTimeline(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetFamilyTimeline(ctx)
	return err
}

// GetHealth converts echo context to params.
func (w *ServerInterfaceWrapper) GetHealth(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetHealth(ctx)
	return err
}

// GetNotifications converts echo context to params.
func (w *ServerInterfaceWrapper) GetNotifications(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetNotifications(ctx)
	return err
}

// GetRecurrenceRules converts echo context to params.
func (w *ServerInterfaceWrapper) GetRecurrenceRules(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetRecurrenceRules(ctx)
	return err
}

// PostRecurrenceRules converts echo context to params.
func (w *ServerInterfaceWrapper) PostRecurrenceRules(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostRecurrenceRules(ctx)
	return err
}

// DeleteRecurrenceRule converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteRecurrenceRule(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "ruleId" -------------
	var ruleId int64

	err = runtime.BindStyledParameterWithOptions("simple", "ruleId", ctx.Param("ruleId"), &ruleId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter ruleId: %s", err))
	}

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteRecurrenceRule(ctx, ruleId)
	return err
}

// PatchRecurrenceRules converts echo context to params.
func (w *ServerInterfaceWrapper) PatchRecurrenceRules(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "ruleId" -------------
	var ruleId int64

	err = runtime.BindStyledParameterWithOptions("simple", "ruleId", ctx.Param("ruleId"), &ruleId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter ruleId: %s", err))
	}

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchRecurrenceRules(ctx, ruleId)
	return err
}

// GetStatsContributions converts echo context to params.
func (w *ServerInterfaceWrapper) GetStatsContributions(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetStatsContributions(ctx)
	return err
}

// GetTaskTemplates converts echo context to params.
func (w *ServerInterfaceWrapper) GetTaskTemplates(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTaskTemplates(ctx)
	return err
}

// GetTasks converts echo context to params.
func (w *ServerInterfaceWrapper) GetTasks(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTasksParams
	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "status", ctx.QueryParams(), &params.Status, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter status: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTasks(ctx, params)
	return err
}

// PostTasks converts echo context to params.
func (w *ServerInterfaceWrapper) PostTasks(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTasks(ctx)
	return err
}

// DeleteTask converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteTask(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskId" -------------
	var taskId int64

	err = runtime.BindStyledParameterWithOptions("simple", "taskId", ctx.Param("taskId"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskId: %s", err))
	}

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteTask(ctx, taskId)
	return err
}

// PatchTask converts echo context to params.
func (w *ServerInterfaceWrapper) PatchTask(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskId" -------------
	var taskId int64

	err = runtime.BindStyledParameterWithOptions("simple", "taskId", ctx.Param("taskId"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskId: %s", err))
	}

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchTask(ctx, taskId)
	return err
}

// PostTasksAppreciation converts echo context to params.
func (w *ServerInterfaceWrapper) PostTasksAppreciation(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskId" -------------
	var taskId int64

	err = runtime.BindStyledParameterWithOptions("simple", "taskId", ctx.Param("taskId"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskId: %s", err))
	}

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTasksAppreciation(ctx, taskId)
	return err
}

// PatchTasksDone converts echo context to params.
func (w *ServerInterfaceWrapper) PatchTasksDone(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskId" -------------
	var taskId int64

	err = runtime.BindStyledParameterWithOptions("simple", "taskId", ctx.Param("taskId"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskId: %s", err))
	}

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchTasksDone(ctx, taskId)
	return err
}

// GetUsersMe converts echo context to params.
func (w *ServerInterfaceWrapper) GetUsersMe(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetUsersMe(ctx)
	return err
}

// PatchUsersMe converts echo context to params.
func (w *ServerInterfaceWrapper) PatchUsersMe(ctx echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchUsersMe(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(options.BaseURL+"/auth/login", wrapper.PostAuthLogin, options.OperationMiddlewares["postAuthLogin"]...)
	router.DELETE(options.BaseURL+"/auth/logout", wrapper.DeleteAuthLogout, options.OperationMiddlewares["deleteAuthLogout"]...)
	router.POST(options.BaseURL+"/auth/refresh", wrapper.PostAuthRefresh, options.OperationMiddlewares["postAuthRefresh"]...)
	router.POST(options.BaseURL+"/auth/register", wrapper.PostAuthRegister, options.OperationMiddlewares["postAuthRegister"]...)
	router.GET(options.BaseURL+"/calendar", wrapper.GetCalendar, options.OperationMiddlewares["getCalendar"]...)
	router.POST(options.BaseURL+"/families", wrapper.PostFamilies, options.OperationMiddlewares["postFamilies"]...)
	router.GET(options.BaseURL+"/families/goals", wrapper.GetFamiliesGoals, options.OperationMiddlewares["getFamiliesGoals"]...)
	router.POST(options.BaseURL+"/families/goals", wrapper.PostFamiliesGoals, options.OperationMiddlewares["postFamiliesGoals"]...)
	router.POST(options.BaseURL+"/families/join", wrapper.PostFamiliesJoin, options.OperationMiddlewares["postFamiliesJoin"]...)
	router.GET(options.BaseURL+"/family/timeline", wrapper.GetFamilyTimeline, options.OperationMiddlewares["getFamilyTimeline"]...)
	router.GET(options.BaseURL+"/health", wrapper.GetHealth, options.OperationMiddlewares["getHealth"]...)
	router.GET(options.BaseURL+"/notifications", wrapper.GetNotifications, options.OperationMiddlewares["getNotifications"]...)
	router.GET(options.BaseURL+"/recurrence-rules", wrapper.GetRecurrenceRules, options.OperationMiddlewares["getRecurrenceRules"]...)
	router.POST(options.BaseURL+"/recurrence-rules", wrapper.PostRecurrenceRules, options.OperationMiddlewares["postRecurrenceRules"]...)
	router.DELETE(options.BaseURL+"/recurrence-rules/:ruleId", wrapper.DeleteRecurrenceRule, options.OperationMiddlewares["deleteRecurrenceRule"]...)
	router.PATCH(options.BaseURL+"/recurrence-rules/:ruleId", wrapper.PatchRecurrenceRules, options.OperationMiddlewares["patchRecurrenceRules"]...)
	router.GET(options.BaseURL+"/stats/contributions", wrapper.GetStatsContributions, options.OperationMiddlewares["getStatsContributions"]...)
	router.GET(options.BaseURL+"/task-templates", wrapper.GetTaskTemplates, options.OperationMiddlewares["getTaskTemplates"]...)
	router.GET(options.BaseURL+"/tasks", wrapper.GetTasks, options.OperationMiddlewares["getTasks"]...)
	router.POST(options.BaseURL+"/tasks", wrapper.PostTasks, options.OperationMiddlewares["postTasks"]...)
	router.DELETE(options.BaseURL+"/tasks/:taskId", wrapper.DeleteTask, options.OperationMiddlewares["deleteTask"]...)
	router.PATCH(options.BaseURL+"/tasks/:taskId", wrapper.PatchTask, options.OperationMiddlewares["patchTask"]...)
	router.POST(options.BaseURL+"/tasks/:taskId/appreciations", wrapper.PostTasksAppreciation, options.OperationMiddlewares["postTasksAppreciation"]...)
	router.PATCH(options.BaseURL+"/tasks/:taskId/done", wrapper.PatchTasksDone, options.OperationMiddlewares["patchTasksDone"]...)
	router.GET(options.BaseURL+"/users/me", wrapper.GetUsersMe, options.OperationMiddlewares["getUsersMe"]...)
	router.PATCH(options.BaseURL+"/users/me", wrapper.PatchUsersMe, options.OperationMiddlewares["patchUsersMe"]...)

}